+++ printer/options.go
@@ -1,0 +1,9 @@
+package printer
+
+// Options is the possible options that the
+// Printer can have
+type Options struct {
+	// ShowIcons toggles display of the
+	// icons on the end Graph
+	ShowIcons bool
+}