+++ printer/printer.go
@@ -9,5 +9,5 @@ import (
 // Printer is an abstraction to Print the graph.Graph
 // in different formats
 type Printer interface {
-	Print(g *graph.Graph, w io.Writer) error
+	Print(g *graph.Graph, opt Options, w io.Writer) error
 }